Refactors AddToMeal function
